test(lsp): delete session.rs convergence property subsumed by pr_session_seeds (#171)
The w10 open/edit/close/restart alphabet is a strict subset of the
session_state_machine operation alphabet (Open/FullEdit/IncrementalEdit/
Close/Restart plus config, baseline, disk-file, folder, and rapid-edit
operations), so the property's coverage is subsumed by pr_session_seeds
and its explicit #53/#55/#45 regression sequences.

Split out first, into session_state_machine.rs, the coverage that was
unique to the file: utf16_contract_holds_across_one_real_lsp_transcript
(hand-declared UTF-16 anchors over BMP/combining/astral/CRLF text) and
close_reopen_publishes_diagnostics_again (#81). The stale proptest
regression seed for the deleted generator goes with it.
