--- crates/ry-lsp/tests/session.proptest-regressions
@@ -1,12 +1,0 @@
-# Seeds for failure cases proptest has generated in the past. It is
-# automatically read and these particular cases re-run before any
-# novel cases are generated.
-#
-# It is recommended to check this file in to source control so that
-# everyone who runs the test benefits from these saved cases.
-#
-# The two seeds recorded against the pre-#65 stateless generator were
-# removed: their value space (unfiltered `Vec<Operation>`) no longer exists,
-# and both encoded protocol violations that are now impossible at generation
-# time.
-cc 188145fde5877a0e85ea15845736f2e13dc8e47bcc16cb13684ff7ba9c43aa5c # shrinks to raw = []
